refactor: a file per directive feature, not one chain of ifs
renderer.ts was 232 lines and thirteen branches deep. It is 52 now, and
builds the context before handing the token to whichever renderer knows
the name.

Split by feature rather than by directive: card.ts holds the card and
everything written inside one, because a title has to know whether the
card gave it an href, and a footer whether to grow an arrow. One file per
name would have put those three apart.

grid-cols-N is looked up after the rest, being a shape rather than a
name, and shares its pattern with registry.ts now instead of carrying a
second one that agreed by luck.

No output changed; the snapshots are untouched.
